Move recovery_upgrader_test to recovery_upgrader_test.rs
