+++ rs/tests/nested/BUILD.bazel
@@ -197,5 +197,7 @@ system_test_nns(
         ":nested",
         "//rs/tests/driver:ic-system-test-driver",
         "@crate_index//:anyhow",
+        "@crate_index//:reqwest",
+        "@crate_index//:slog",
     ],
 )
