+++ rs/tests/nested/src/lib.rs
@@ -1,5 +1,3 @@
-use std::time::Duration;
-
 use anyhow::bail;
 use ic_system_test_driver::{
     driver::{
@@ -10,14 +8,10 @@ use ic_system_test_driver::{
     retry_with_msg,
     util::block_on,
 };
-use reqwest::Client;
 use slog::info;
 
 pub mod util;
-use util::{
-    NODE_REGISTRATION_BACKOFF, NODE_REGISTRATION_TIMEOUT, get_host_boot_id,
-    setup_ic_infrastructure, wait_for_guest_version,
-};
+use util::{NODE_REGISTRATION_BACKOFF, NODE_REGISTRATION_TIMEOUT, setup_ic_infrastructure};
 
 pub const HOST_VM_NAME: &str = "host-1";
 
@@ -102,119 +96,4 @@ pub fn registration(env: TestEnv) {
         }
     ).unwrap();
     info!(logger, "All {n} nodes successfully came up and registered.");
-}
-
-/// Test the guestos-recovery-upgrader component: tests upgrading the GuestOS
-/// from the HostOS based on injected version/hash boot parameters.
-pub fn recovery_upgrader_test(env: TestEnv) {
-    let logger = env.logger();
-
-    let host = env
-        .get_nested_vm(HOST_VM_NAME)
-        .expect("Unable to find HostOS node.");
-    let guest_ipv6 = host
-        .get_nested_network()
-        .expect("Unable to get nested network")
-        .guest_ip;
-
-    block_on(async {
-        let client = Client::builder()
-            .danger_accept_invalid_certs(true)
-            .build()
-            .expect("Failed to build HTTP client");
-
-        let original_version = wait_for_guest_version(
-            &client,
-            &guest_ipv6,
-            &logger,
-            Duration::from_secs(10 * 60), // long wait for setupOS to install
-            Duration::from_secs(5),
-        )
-        .await
-        .expect("guest didn't come up as expected");
-
-        info!(
-            logger,
-            "Retrieving the current boot ID from the host before we update boot_args so we can determine when it rebooted..."
-        );
-        let host_boot_id_pre_reboot = get_host_boot_id(&host);
-        info!(
-            logger,
-            "Host boot ID pre reboot: '{}'", host_boot_id_pre_reboot
-        );
-
-        info!(logger, "Checking current boot_args file content");
-        let current_boot_args = host
-            .block_on_bash_script("cat /boot/boot_args")
-            .expect("Failed to read /boot/boot_args file");
-        info!(logger, "Current boot_args content:\n{}", current_boot_args);
-
-        let target_version = get_guestos_update_img_version();
-        let target_version_hash = get_guestos_update_img_sha256();
-
-        info!(
-            logger,
-            "Using target version: {} and version-hash: {}", target_version, target_version_hash
-        );
-
-        info!(
-            logger,
-            "Remounting /boot as read-write and updating boot_args file"
-        );
-        let boot_args_command = format!(
-            "sudo mount -o remount,rw /boot && sudo sed -i 's/\\(BOOT_ARGS_A=\".*\\)enforcing=0\"/\\1enforcing=0 recovery=1 version={target_version} version-hash={target_version_hash}\"/' /boot/boot_args && sudo mount -o remount,ro /boot"
-        );
-        host.block_on_bash_script(&boot_args_command)
-            .expect("Failed to update boot_args file");
-        info!(logger, "Boot_args file updated successfully.");
-
-        info!(logger, "Verifying boot_args file contents");
-        let updated_boot_args = host
-            .block_on_bash_script("cat /boot/boot_args")
-            .expect("Failed to read updated /boot/boot_args file");
-        info!(logger, "Updated boot_args content:\n{}", updated_boot_args);
-
-        info!(logger, "Rebooting the host");
-        host.block_on_bash_script("sudo reboot")
-            .expect("Failed to send reboot command (connection may be terminated by reboot)");
-
-        info!(logger, "Waiting for host to reboot...");
-
-        retry_with_msg!(
-            format!(
-                "Waiting until the host's boot ID changes from its pre reboot value of '{}'",
-                host_boot_id_pre_reboot
-            ),
-            logger.clone(),
-            Duration::from_secs(5 * 60),
-            Duration::from_secs(5),
-            || {
-                let host_boot_id = get_host_boot_id(&host);
-                if host_boot_id != host_boot_id_pre_reboot {
-                    info!(
-                        logger,
-                        "Host boot ID changed from '{}' to '{}'",
-                        host_boot_id_pre_reboot,
-                        host_boot_id
-                    );
-                    Ok(())
-                } else {
-                    bail!("Host boot ID is still '{}'", host_boot_id_pre_reboot)
-                }
-            }
-        )
-        .unwrap();
-
-        let new_version = wait_for_guest_version(
-            &client,
-            &guest_ipv6,
-            &logger,
-            Duration::from_secs(5 * 60),
-            Duration::from_secs(5),
-        )
-        .await
-        .expect("guest didn't come up as expected");
-
-        assert!(new_version != original_version);
-    });
 }
